requirements.txt
python-telegram-bot>=20.0
openai>=0.27.0
gtts>=2.3.0
python-dotenv>=1.0.0
